+++ test/config/config.js
@@ -1,3 +1,5 @@
+'use strict';
+
 module.exports = {
   username: process.env.SEQ_USER || 'sequelize_test',
   password: process.env.SEQ_PW   || null,