+++ test/support.js
@@ -5,9 +5,14 @@ var fs = require('fs'),
 	_ = require('lodash'),
 	Sequelize = require('sequelize'),
 	DataTypes = require(__dirname + '/../node_modules/sequelize/lib/data-types'),
-	Config = require(__dirname + '/config/config');
+	Config = require(__dirname + '/config/config'),
+	chai = require('chai'),
+	expect = chai.expect,
+	chaiAsPromised = require('chai-as-promised');
 
 require('../lib/index')(Sequelize);
+
+chai.use(chaiAsPromised);
 
 // Make sure errors get thrown when testing
 Sequelize.Promise.onPossiblyUnhandledRejection(function(e, promise) {
@@ -76,7 +81,7 @@ var Support = {
 
 		var sequelizeOptions = _.defaults(options, {
 			host: options.host || config.host,
-			logging: false,
+			logging: (process.env.SEQ_LOG ? console.log : false),
 			dialect: options.dialect,
 			port: options.port || process.env.SEQ_PORT || config.port,
 			pool: config.pool,
@@ -105,16 +110,12 @@ var Support = {
 			.getQueryInterface()
 			.dropAllTables()
 			.then(function() {
-				sequelize.daoFactoryManager.daos = [];
+				sequelize.modelManager.daos = [];
+				sequelize.models = {};
+
 				return sequelize
 					.getQueryInterface()
-					.dropAllEnums()
-					.catch(function (err) {
-						console.log('Error in support.clearDatabase() dropAllEnums() :: ', err);
-					});
-			})
-			.catch(function(err) {
-				console.log('Error in support.clearDatabase() dropAllTables() :: ', err);
+					.dropAllEnums();
 			});
 	},
 
@@ -184,25 +185,28 @@ var Support = {
 			url = config.dialect + '://' + credentials + '@' + dbConfig.host + ':' + dbConfig.port + '/' + dbConfig.database;
 		}
 		return url;
+	},
+
+	expectsql: function(query, expectations) {
+		var expectation = expectations[Support.sequelize.dialect.name];
+
+		if (!expectation && Support.sequelize.dialect.name === 'mariadb') {
+			expectation = expectations['mysql'];
+		}
+
+		if (!expectation) {
+			expectation = expectations['default']
+				.replace(/\[/g, Support.sequelize.dialect.TICK_CHAR_LEFT)
+				.replace(/\]/g, Support.sequelize.dialect.TICK_CHAR_RIGHT);
+		}
+
+		expect(query).to.equal(expectation);
 	}
 };
 
-var sequelize = Support.createSequelizeInstance();
-//
-// For Postgres' HSTORE functionality and to properly execute it's commands we'll need this...
-before(function() {
-	var dialect = Support.getTestDialect();
-	if (dialect !== "postgres" && dialect !== "postgres-native") {
-		return;
-	}
-
-	return sequelize.query('CREATE EXTENSION IF NOT EXISTS hstore', null, {raw: true});
-});
-
 beforeEach(function() {
-	this.sequelize = sequelize;
-
-	return Support.clearDatabase(this.sequelize);
+	this.sequelize = Support.sequelize;
 });
 
+Support.sequelize = Support.createSequelizeInstance();
 module.exports = Support;