+++ changelog.md
@@ -142,6 +142,7 @@ Now supports all Sequelize dialects.
 
 * Update sequelize dependency to v2.0.0+
 * Update dev dependencies in line with sequelize v2.0.5
+* Update test support files in line with sequelize v2.0.5
 * Code tidy in test/support.js
 * Travis runs tests against node 0.10 and 0.12
 * Correct typo in changelog